Now a Geom3D.SymmetricShape instance can be saved as .py (by writing its repr)
The SymmetricShape should reflect the Scene_Orbit user interface.
Was trying to write doc for this, with an example cube, but the example didn't
work. E.g. missing subgroups in S4xI. Then I felt again the need to be able to
import shapes into Scene_Orbit. It would be best to import them as .py (instead
of .off)
Now this is possible and it can be read and interpreted by importing (or by exec
in dict)
Alse added D8D4, D4C4, and C4xI as subgroup of S4xI