Added some missing subgroups
TODO: add these missing subgroups to D2nDn and DnxI realiseSubgroups.
TODO:
Something still wrong for colouring the model:
E.g. for S4xI | D2C2 you don't know the orientation of the stabiliser:
if the principle axis is an 4-fold or 2-foold axis. The realisesubgroups just
generates all and if the stabiliser is at a 2-fold axis you get some strange
colourings.