Moved out orbit part into separate lib
Tested ok. This is as a preparation for handling 2nd order colouring. For the
colouring up to now it holds that a symmetry operation of the final model will
either turn one colour into another, or it will map one of the colours onto
itself. This property doesn't hold for using 5 colours for a icosahedron (where
faces of one colour have an A4 symmetry): a reflection true a triangle will map
the triangle on itself but not the other triangles of the same colour. To handle
these cases I will break this out: then I can predefine which ones are allowed
(if I want that)