Added support for alternative colourings: buggy..
Added support for direct parents. For a mised group this is the biggest subgroup
that only consists of direct isometries.
This is used to get alternative higher order stabilisers. These alternatives
arise when both the final group and the stabiliser are mixed. Then instead one
can look at the direct parents of these. This can be used for colouring an
icosahedron with 5 colours.

However: it doesn't work so well yet.