folded heptagons: change triangle fill GUI
This patch switches to a choice list for the triangle fill alternatives.
The previous interface looked simpler for the user, but was pretty
complicated to implement. I ran into problems with the predefined
positions. Then it became impossible to see for a user which triangle
alternatives had solutions. E.g. I had to enable the '1 loose' checkbox,
if only one '1 loose' alternative had solutions.

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>