roots: fixes for splitting solutions
This fixes an error in splitting solutions: fold1 and 2 needs to be
compared too, since solutions exists for which only the two different
fold2 differ. These would be written as a reflective solution, having
only 5 settings. The next time these are read they are invalid and are
filtered out (while they probably are found again as well)

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>