roots: fix precision problems and dynamic solution
This fixes problems with precision:
- before this patch doublets were not recognised so well.
- dynamic solutions were not regnised so well
This patch will write the solutions with one more digit than the
required precision so that the calculated edge length has the required
precision, this means that all solution files will be written with less
digits, since before all known digits were written.

This also fixes the latest update for filtering out a dynamic solution.

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>