Geom3D, Orbit: fix reprs and interpret compound
This fixes the orbit scene so it can interpret the Compound shapes.

This fixes the reprs for the Compound shape, so the same shape is
obtained by re-interpreting, replacing the flattened shape.
It fixes the repr for the Isometric shape by applying the isometries.
It also removes the getDome and toOffStr for that shape, since these
function are the same as for the parent.

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>