roots: make twisted solutions S4A4 solutions
This creates a new group of solutions: S4A4 and moves the twisted edge
alternative into that group. The twisted solutions (in the way they are
defined today) cannot be combined with other edge alternatives (like in
A4 for instance) and the only angle they make sense is by turning one
eight turn (otherwise the hexagon isn't flat).

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>