roots: make it possible to run once through all
This adds a new command line option '-s' that makes it possible to run
through all set-ups once and then exits, instead of looping
indefenitely. This can be handy, eg, to increase or decrease precision
of all solutions (in combination with -i 1). Updates TODO as well,
removes the old one and adds a new one.

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>