roots: fix more doublets problems
This introduces a property eq_margin that is used for comparing
solutions. It must be bigger than the precision, since settings that
vary a bit more than the precision can still produce edge lengths that
are within the precision limit. Currently it is 10 times bigger. Now
this margin is used in other comparisons.

This fixes problems with:
  $ python roots.batch.py -p 5 -i 100  -l 0 -f 2  S4A4

Updates TODO as well

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>