roots: change precision again
This interprets the precision the other way around: the specified
precision is the precision that is used to write the values with. It is
also the precision of the edge lengths, though this would mean that the
actual precision of the setup vector is less, which means that a higher
margin should be used to see if they can be interpreted as equal
solutions.

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>