roots: prevent unecessary writes a solutions
This will prevent writing a new solution value if the new set-up differs
less than the required margin (which is one digit less than the
precision that is used to write solutions).

It also catches exception in the threads that are started and will exit
the parent if there is a child exception.

Signed-off-by: Marcel Tunnissen <marcelteun@gmail.com>