Fix subgroups for Cn
This fixes problems with the colours in the Orbit scene when using the
Cn symmetry with specific n. These problems were caused because
subgroups were only define when using the specific meta class. Besides
that the realisation of the subgroups were needed. In addition all
subgroups are defined now instead of just the trivial ones.

Similar fixes are needed for the other cyclic and dihedral groups.