Move code to set define a rotation to own GUI class
The code in Scene_Orbit that implements a GUI that can be used to define
a rotation around an axis is move into its own class in GeomGui that
Scene_Orbit can use. The code is slightly improved so the pressing ENTER
or TAB in the field that defines the angle will set the angle.