orbit.py: fix assert subGroup.order!=0, cyclic
This assertion error is obtained when using extended cylcic and
dihedral. This patch fixes the error for cyclic groups: CnxI, C2nCn, and
DnCn. The problem is that the subgroups weren't defined: some TODOs
needed to be fixed. The weren't done, since it was a bit complicated. I
added unit tests for this.

Note that there is still a problem. After changing the axis of order of
C2nCn, CnxI or DnCn, the colour setup needs to be recalculated.