Heptagons: show more alternatives
Note sure why a removed many options before, but probably because the I
reasoned that the options A - B wasn't essentially different from B - A
(I should be able to find the similar solutions) because you go from one
order 3 symmetry axis to another order 3 symmetry axis. However one side
of the axis doesn't look the same as the other side of the axis, for
instance this is the case with a tetrahedron: one side of the axis has a
triangle, the other side a vertex. Anyway I found solution for A - B,
while I didn't find for B - A.