scene orbit: fix problem with new wxWidgets version
Fixes the following problem in the log when repeatedly applying a
symmetry:
Traceback (most recent call last):
  File "<string>", line 422, in on_apply_sym
  File "<string>", line 311, in add_col_gui
  File "<string>", line 471, in on_no_of_col_select
RuntimeError: wrapped C/C++ object of type BoxSizer has been deleted

Before an AttributeError was raised (and handled) nowadays only a C/C++
error is generated. Solution is to reset the the sizer to None so that
the AttributeError is caught on a higher level anyway.