Start to add support for saving objects as JSON
The goal is to get rid of loading modesl by interpreting Python.
This makes it easier to convert to a loadable JSON file.
Also cleanup IsometricShape by removing unnecessary split of isometries.
Rename some identifiers to use snake case.
Remove a lot of debug prints from Geom3D.
Get rid of legacy parameter unfoldOrbit

STATUS
The objects that can be saved are SimpleShape and CompoundShape,
IsometricShape.
The unit test for those work
For IsometricShape the order of the isometries shouldn't matter the test
should perhaps be more robust (or allow and check if it changes..)
For SymmetricShape the isometries chosem from the quotient set might be different alltogether.

CONTINUE with repr_dict for SymmetricShape
Run
python unittest/tst_geom3d.py
And fix errors.