Improve consistency for face colours after orbits
There seem to be a problem with faces having a mixture of dark and
bright colours. This might be related to orbits choosing indirect
isometries. Try not to use these. Also warn when orbiting doesn't
lead to the right amount of operations (due to lack of precision)

There is still a problem that faces become transparent sometimes
after orbiting them.