Fix github issue no. 7
The issue was that when JSON files are imported and the chooses to see
the stabilizer symmetry it always shows "E". When the users chooses
"Show Stabilizer" the list with stabilizers is always regenerated
to keep this list up to date. This is how the interface works: when the
user chooses a new final symmetry, the list of stabilizers is hidden and
the list is regenerated when the users choses to see them.

The easiest fix is to show the list when importing and then set the
stabilizer. Another option would be to regenerate the list only when a
new final symmetry is selected (or when the value 'n' for cyclic or
dihedral symmetries is updated). This is however much more work.