Fix that some JSON models become transparent
This fixes github issue #5. This issue was seen with certain models
using concave faces when these were imported to scene_orbit as JSON
files. If an alpha channel=1 was used for the colour then the problem
was seen, but not if no alpha channel was used in the JSON file.

The faces seem to become transparent sometimes due to the part with the
stencil buffer, where the depth buffer is disabled. This is done
differently for colours that have an alpha channel or not. I tried to go
back in history to check why this was done like this and why certain
code was commented out, but this has been like this from the beginning
(of the git history). The history from the previous versioning system
(RCS) is lost. I chose to handle the alpha case in the same way as the
opaque case. I checked with using alpha for
MW107.bas.10_3.A5xI.D5xI.json and this seemed to work as well.

I also added support for blending specifically and make sure alpha value
use the right domain.