Reorganise float handling
Create a singleton class to compare floats with certain margin and use
it to print the right amount of decimals. With default precision of 12
decimals, use a construction like this for standard floats:
    with geomtypes.FloatHandler(precision=7) as fh:
        if fh.eq(0, 0.00000000000001):
            print("equal to zero")
or create a variable of the type RoundedFloat
    a = geomtypes.RoundedFloat(0.00000000000001)
    if a == 0:
        print("equal to zero")

Will print "equal to zero"

Improved to_postscript (since it was updated). Now faces that are in one
plane are printed together in one picture.