geom_3d: simplify face/line intersection algorithm
This is the first step into simplifying and improvig the algorithm where
a line intersects a face. It uses the new algorithm in geom_2d, which
works in more cases (where a line meets in a vertex).

The long unit test takes around the same time. The result for MW119 is
different in ASCII but comparing the result in gv doesn't show any
differences. Apparently some of the intersections where shown multiple
times before. This was verified by hand for some.