geom_3d: step 2 improve and simplify intersecting
Instead of the old misplaced and ugly method intersect_with_face that is full of bugs,
use the new and more trusted intersect_polygon_get_factors.

The unit test for MW119 contains a new version. Some of these are due to
precision, for others these are related to extra intersections being
removed. This was verified manually for some cases. All differences
occur on page 7 and gv shows no difference in the result.