Increase output float precision
Add one more digit. This is useful for the JSON files for uniform
polyhedra. When the 5-fold axis for A5 is saved with the precision of 12
and the file is loaded into the orbit scene then many infos are show
with the message the isometries are treated as the same, while their
hashes aren't. Increasing the float precision with one will fix that.