Orbit scene: fix segmentation fault
Segmentation faults occurred when choosing lower order symmetry. E.g.
after choosing and applying a final symmetry of A5, then a segmentation
fault occurred as soon as the final symmetry E was selected from the
list.

I think that the problem is that sizers shouldn't be destroyed by the
destroy function.