Add support for cyclic cube compoundsa: CnxI
Add support for loading CnxI JSON in orbit scene
Add support for loading specific CnxI, e.g. C4xI JSON file, while list
of groups only lists the general CnxI class. In this case we need to
check whether the class C4xI is inherited from CnxI.

In isometry: make sure the generator is an attribute
This is important when exporting in JSON format and then loading the
JSON in the orbit scene. If there is no generator, then there is not
rotation axis (n-fold axis). In that case reconstruting fails.

Use new parameter name for generating compounds and rename the parameter
'n' is called 'n_domain' now.

Save JSON for compounds of objects with cube symmetry for CnxI symmetry

Make sure the axis and angle is saved for cyclic symmetries.