Tetrahedron compounds: update requirement of n etc
The handling of no_of parameter is updated so that it certain groups are
isomorph, e.g. C2 and D1, then only one is used to generate a compound.

Also update the the n-fold axis for C2C1 so that the relection is shown
when looking down the z-axis. Since the 1-fold axis doesn't show
anything.