Fix issue 24: missing line in PS for 4_D4xI_D2C2
For the compounds of tetrahedra with the notation 4 | D4xI / D2C2 the
PostScript file was missing two lines. This was because vertices were
projected on the horizontal plane orthogonally the simplify the
calculation, but the face was orthogonal to the horizontal plane and all
vertices that should form a polygon ended up on one line. Hence no
intersection line was found for that polygon. The simplified solution
cannot be used in this case.

Also fix some code formatting, comments and doc strings on the fly

TODO: add test case

Solves: Issue 24