Orbit scene: make sure that stabiliser list is updated
I tried with D2nDn: with order 12, but I still had to regenerate.

When the order parameter of the cyclic and dihedral symmetries is
updated, then the list of stabilisers need to be updated by regenerating
them. This is done by watching the order parameter and call a call-back
method to the scene (unless the input value is deleted)