Add test for isometry and rewrite existing
Add a test that test subgroups in practice by obtaining an object for
the extended dihedral groups and checking for different possible
subgroups whether all isometries are part of the main group. Compare
this with the defined subgroups of that one.

Some errors were found and fixed. This meant fixing the results for the
existing test as well, which was reorgnised to become more Pythonic.

Also added some more explanation about the dihedral extensions