Added an new event for handling order
The text event didn't always come through, especially when typing more
than one number in the order field, e.g. 12.

This adds an event for each character typed with a delay so that the
right value is obtained when doing gui.GetValue otherwise I think it
might depend on the order of event handling whether the right value is
obtained.

Also add messages in case the number is high, e.g. D2nDn with n=300
takes a long time and in that case the GUI hangs for a while. Now this
at leasts shows a message in the status bar, which is updated when
ready.