Add method to get concave points
Add Face class with method to convert e.g. a pentagram consisting of
five vertices to a polygon with 10 vertices, for which the five
"vertices" where the edges meet are added.

What this method does is that it takes follows the outer edges and hence
there will not be any holes left. This method shouldn't be used if these
holes are needed. In that case one should use the stencil buffer.

The Face class assumes that vertices come in the order in which they are
connected. This means that the Face class cannot represent a compound of
Faces e.g.

Unit tests are added for several cases.