Fix isuuse with outline for certain polygrams
The method for outline of polygrams didn't work for the face of a noble
stellation of the RTC, number 6 on Miraheze. Apparently one cannot rely
on the normal automatically. For this face the direction of the normal
depends on where you start.

This patch introduces a method flip_normal to handle this and it adds a
test case with the failing polygram.

In the future one could think of something that works automatically. I
tried checking whether all vertices were visited, but then test case
'hole-or-no' fails. I think one should test whether the resulting
polygram is in general clockwise or counter-clockwise.