Fix wx.Slider problems with input type
This is a quick fix to get sliders to show at least.

It seems the the Slider doesn't support sending float anymore. For now
just convert the values to integer, but it might be that the slider
doesn't work properly anymore, e.g. when the min. value was 0.1 and the
max. value was 0.9. It is probably better just to introduce some proper
scaling and wrap the Slider with my own type or something like that.