Fix issue with save OFF after changing colours
After updating the colours using Ctrl+f the new colours weren't saved.
This is solved now.

The problem was related to the fact that the colours were updated for
each simple shape in a compound of shapes and the attribute merge_needed
wasn't updated for the compound shape. As a consequence the merge wasn't
performed when exporting the shape to OFF. Since there is already a
call-back to update_shape_cols from each faces-tab, which is per
SimpleShape, this was easy to solve.

Solves: github #26